Write events to a binary buffer
Before, we allocated a heap-value for each event, which made work for
the GC. Now, we pre-allocate a Bigarray and write the events to that in
a binary format (described in the "metadata" file, which can be used
with babeltrace to parse the logs).

Times are now 64-bit ints and record the monotonic timer rather than
using gettimeofday. At some point, we should use CPU counters here.
Thread IDs are now 64-bit, as OCaml ints might wrap on 32-bit systems.

TODO:
- trace header
- minios C stubs